split monolith lint file into more reasonably sized files
